reg update before hub

+++ wp-content/plugins/load-more-products-for-woocommerce/includes/main.php
@@ -6,9 +6,11 @@ if( ! class_exists('BeRocket_selector_wizard_woocommerce') ) {
             add_action('BeRocket_wizard_start', array($this, 'import_products'));
             add_action('BeRocket_wizard_javascript', array($this, 'javascript'));
             add_action('BeRocket_wizard_end', array($this, 'remove_products'));
+            add_action('BeRocket_wizard_ended_check', array($this, 'remove_products_ended'));
             add_filter('BeRocket_wizard_category_link', array($this, 'category_link'));
             add_action( 'wp_ajax_berocket_wizard_selector_start', array( $this, 'wizard_selector_start' ) );
             add_action( 'wp_ajax_berocket_wizard_selector_end', array( $this, 'remove_products' ) );
+            add_action( 'wp_ajax_berocket_wizard_selector_ended', array( $this, 'remove_products_ended_check' ) );
             $status = get_option('BeRocket_selector_wizard_status');
             if( $status == 'started' ) {
                 add_filter('loop_shop_per_page', array($this, 'loop_shop_per_page'), 999999999999999 );
@@ -35,6 +37,8 @@ if( ! class_exists('BeRocket_selector_wizard_woocommerce') ) {
                     'creating_products' => __('Creating products', 'BeRocket_domain'),
                     'getting_selectors' => __('Gettings selectors', 'BeRocket_domain'),
                     'removing_products' => __('Removing products', 'BeRocket_domain'),
+                    'stopping_text' => __('Stopping. Wait', 'BeRocket_domain'),
+                    'stop_text' => __('Stop', 'BeRocket_domain'),
                     'error' => __('Error:', 'BeRocket_domain'),
                     'ajaxurl' => admin_url('admin-ajax.php')
                 )
@@ -130,9 +134,7 @@ if( ! class_exists('BeRocket_selector_wizard_woocommerce') ) {
             }
         }
         
-        function remove_products() {
-            remove_filter('loop_shop_per_page', array($this, 'loop_shop_per_page'), 999999999999999 );
-            remove_action('pre_get_posts', array($this, 'products_per_page'), 999999999999999);
+        public function remove_products_ended() {
             add_action('pre_get_posts', array($this, 'products_per_page_more'), 999999999999999);
             $args = array("post_type" => "product", "s" => 'BeRocketSelectorsTest', 'posts_per_page'   => 100);
             $query = get_posts( $args );
@@ -143,13 +145,30 @@ if( ! class_exists('BeRocket_selector_wizard_woocommerce') ) {
                 }
             }
             $term = get_term_by('name', 'BeRocketSelectors', 'product_cat');
-            wp_delete_term($term->term_id, $term->taxonomy);
+            if( $term !== FALSE ) {
+                wp_delete_term($term->term_id, $term->taxonomy);
+            }
+            remove_action('pre_get_posts', array($this, 'products_per_page_more'), 999999999999999);
+        }
+        function remove_products_ended_check() {
+            $status = get_option('BeRocket_selector_wizard_status');
+            if( $status == 'ended' ) {
+                do_action('BeRocket_wizard_ended_check');
+            }
+            wp_die();
+        }
+        function remove_products() {
+            remove_filter('loop_shop_per_page', array($this, 'loop_shop_per_page'), 999999999999999 );
+            remove_action('pre_get_posts', array($this, 'products_per_page'), 999999999999999);
+            $this->remove_products_ended();
             update_option('BeRocket_selector_wizard_status', 'ended');
             wp_die();
         }
     }
     new BeRocket_selector_wizard_woocommerce();
     function BeRocket_wizard_generate_autoselectors($input_classes = array(), $js_functions = array()) {
+        $status = get_option('BeRocket_selector_wizard_status');
+        $status = $status == 'started';
         $js_functions = array_merge(
             array(
                 'any' => '',
@@ -167,12 +186,22 @@ if( ! class_exists('BeRocket_selector_wizard_woocommerce') ) {
                 'result_count' => '',
             ), $input_classes);
         $html = '<div class="berocket_wizard_autoselectors" data-functions=\'' . json_encode($js_functions) . '\' data-inputs=\'' . json_encode($input_classes) . '\'>
-            <button class="berocket_autoselector button" type="button">Auto-Selectors</button>
-            <span class="berocket_autoselect_spin" style="display:none;"><i class="fa fa-spinner fa-pulse fa-3x fa-fw"></i></span>
+            '.($status ? '<div class="berocket_selectors_was_runned"><strong style="color: red;">'.__('Script was runned, but page closed until end. Please stop it to prevent any problems on your site', 'BeRocket_domain').'</strong></div>' : '').'
+            <button class="berocket_autoselector button" type="button"'.($status ? ' disabled' : '').'>Auto-Selectors</button>
+            <span class="berocket_autoselect_spin"'.($status ? '' : 'style="display:none;').'"><i class="fa fa-spinner fa-pulse fa-3x fa-fw"></i><button class="berocket_autoselector_stop button" type="button">Stop</button></span>
             <span class="berocket_autoselect_ready" style="display:none;"><i class="fa fa-check fa-3x fa-fw"></i></span>
             <span class="berocket_autoselect_error" style="display:none;"><i class="fa fa-times fa-3x fa-fw"></i></span>
             <div class="berocket_autoselector_load" style="display:none;"><div class="berocket_line"></div><div class="berocket_autoselector_action"></div></div>
-        </div>';
+            <h4>' . __('Steps:', 'BeRocket_domain') . '</h4>
+            <ol>
+                <li><span>' . __('Creating products', 'BeRocket_domain') . '</span> <i class="fa"></i></li>
+                <li><span>' . __('Gettings selectors', 'BeRocket_domain') . '</span> <i class="fa"></i></li>
+                <li><span>' . __('Removing products', 'BeRocket_domain') . '</span> <i class="fa"></i></li>
+            </ol>
+        </div>
+        <script>jQuery(document).ready(function(){
+            jQuery.get(berocket_wizard_autoselect.ajaxurl, {action:"berocket_wizard_selector_ended"});
+        });</script>';
         return $html;
     }
 }
